hospital-ai-service/requirements.txt
flask==2.3.2
langchain-google-genai==0.0.7
python-dotenv==1.0.0